Make swerve auton builder to make autons easier to make (#12)
Creates a SpartanSwerveAutonBuilder, which is a derivative of SwerveAutonBuilder from PathPlanner. It includes the ability to have gyro yaw be set to the heading of the ending point, allowing drivers to not have to reset the gyro going into teleop. They can just drive forward, and just re orientate their heading.